+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.ReservationTimeRequest;
-import roomescape.dto.ReservationTimeResponse;
-import roomescape.dto.ReservationTimesWithTotalPageResponse;
+import roomescape.dto.request.ReservationTimeRequest;
+import roomescape.dto.request.ReservationTimesWithTotalPageResponse;
+import roomescape.dto.response.ReservationTimeResponse;
 import roomescape.service.ReservationTimeService;
 
 @RestController
