+++ src/main/java/roomescape/controller/ThemeController.java
@@ -14,10 +14,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.AvailableReservationResponse;
-import roomescape.dto.ThemeRequest;
-import roomescape.dto.ThemeResponse;
-import roomescape.dto.ThemesWithTotalPageRequest;
+import roomescape.dto.request.ThemeRequest;
+import roomescape.dto.request.ThemesWithTotalPageRequest;
+import roomescape.dto.response.AvailableReservationResponse;
+import roomescape.dto.response.ThemeResponse;
 import roomescape.service.ThemeService;
 
 @RestController
