+++ src/main/java/roomescape/dto/request/ReservationTimesWithTotalPageResponse.java
@@ -1,6 +1,7 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 import java.util.List;
+import roomescape.dto.response.ReservationTimeResponse;
 
 public record ReservationTimesWithTotalPageResponse(int totalPages, List<ReservationTimeResponse> reservationTimes) {
 }
