+++ src/main/java/roomescape/dto/request/ThemesWithTotalPageRequest.java
@@ -1,6 +1,7 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 import java.util.List;
+import roomescape.dto.response.ThemeResponse;
 
 public record ThemesWithTotalPageRequest(int totalPages, List<ThemeResponse> themes) {
 }
