+++ src/main/java/roomescape/dto/response/ReservationsWithTotalPageResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.response;
 
 import java.util.List;
