+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -4,9 +4,9 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
-import roomescape.dto.ReservationTimeRequest;
-import roomescape.dto.ReservationTimeResponse;
-import roomescape.dto.ReservationTimesWithTotalPageResponse;
+import roomescape.dto.request.ReservationTimeRequest;
+import roomescape.dto.request.ReservationTimesWithTotalPageResponse;
+import roomescape.dto.response.ReservationTimeResponse;
 import roomescape.model.ReservationTime;
 
 @Service
