+++ src/main/java/roomescape/service/ThemeService.java
@@ -8,10 +8,10 @@ import org.springframework.stereotype.Service;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.dao.ThemeDao;
-import roomescape.dto.AvailableReservationResponse;
-import roomescape.dto.ThemeRequest;
-import roomescape.dto.ThemeResponse;
-import roomescape.dto.ThemesWithTotalPageRequest;
+import roomescape.dto.request.ThemeRequest;
+import roomescape.dto.request.ThemesWithTotalPageRequest;
+import roomescape.dto.response.AvailableReservationResponse;
+import roomescape.dto.response.ThemeResponse;
 import roomescape.model.Reservation;
 import roomescape.model.ReservationTime;
 import roomescape.model.Theme;
