+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -15,9 +15,9 @@ import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.dao.ThemeDao;
-import roomescape.dto.ReservationRequest;
-import roomescape.dto.ReservationResponse;
-import roomescape.dto.ReservationsWithTotalPageResponse;
+import roomescape.dto.request.ReservationRequest;
+import roomescape.dto.response.ReservationResponse;
+import roomescape.dto.response.ReservationsWithTotalPageResponse;
 import roomescape.model.ReservationTime;
 import roomescape.model.Theme;
